+++ tasq/cli/main.py
@@ -13,8 +13,9 @@ from ..settings import get_config
 
 def get_parser():
     parser = argparse.ArgumentParser(description='Tasq CLI commands')
+    parser.add_argument('subcommand')
     parser.add_argument('-f', action='store')
-    parser.add_argument('--worker', action='store_true')
+    parser.add_argument('--secure', '-s', action='store_true')
     parser.add_argument('--workers', nargs='*')
     parser.add_argument('--random', action='store')
     parser.add_argument('--debug', action='store_true')
@@ -23,19 +24,19 @@ def get_parser():
     return parser
 
 
-def start_worker(host, port, debug):
+def start_worker(host, port, debug, sign_data):
     from tasq.remote.master import Master
-    master = Master(host, port, port + 1, debug=debug)
+    master = Master(host, port, port + 1, debug=debug, sign_data=sign_data)
     master.serve_forever()
 
 
-def start_workers(workers, debug):
+def start_workers(workers, debug, sign_data):
     from tasq.remote.master import Masters
     masters = Masters(workers, debug=debug)
     masters.start_procs()
 
 
-def start_random_workers(host, num_workers, debug):
+def start_random_workers(host, num_workers, debug, sign_data):
     import random
     from tasq.remote.master import Masters
     workers_set = set()
@@ -64,8 +65,11 @@ def main():
         conf = get_config(path=args.f)
     host, port = conf['host'], conf['port']
     debug = conf['debug']
+    sign_data = conf['sign_data']
     if args.debug:
         debug = True
+    if args.secure:
+        sign_data = True
     if args.workers:
         try:
             pairs = conf['workers']
@@ -77,12 +81,12 @@ def main():
                 workers = _translate_peers(pairs)
         else:
             workers = _translate_peers(args.workers or conf['workers'])
-        start_workers(workers, debug)
-    if args.worker:
+        start_workers(workers, debug, sign_data)
+    if args.subcommand == 'worker':
         if args.addr:
             host = args.addr
         if args.port:
             port = int(args.port)
-        start_worker(host, port, debug)
+        start_worker(host, port, debug, sign_data)
     elif args.random:
-        start_random_workers(host, int(args.random), debug)
+        start_random_workers(host, int(args.random), debug, sign_data)